apps/ml-engine/requirements.txt
pandas
numpy
yfinance
psycopg2-binary
sqlalchemy
google-generativeai
requests
httpx
fastapi
uvicorn
python-dotenv
pydantic
scikit-learn
ta
scipy
apscheduler
redis
pytest
boto3
